feat: added versioning to schemas
